% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_f}
\alias{clean_f}
\title{clean_f}
\usage{
clean_f(obj)
}
\arguments{
\item{obj}{A function.}
}
\value{
A list that contains \code{ret} and \code{equations}. \code{ret}
provides the return argument from the function and \code{equations}
provides a character vector of the equations with spacing removed.
}
\description{
Extract and clean ODE equations in user defined function
}
