% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_prog.R
\name{read_stan_file}
\alias{read_stan_file}
\title{read_stan_file}
\usage{
read_stan_file(has_events, integrator, sampling)
}
\arguments{
\item{has_events}{A logical value that indicates whether or not the model contains events.}

\item{integrator}{The type of integrator to use. One of either \code{"rk45"} or \code{"bdf"}.}

\item{sampling}{Logical variable indicating whether the user wants to perform sampling.}
}
\description{
Return the path of the desired Stan program
}
