% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{student_t}
\alias{student_t}
\title{Student-t Distrubtion}
\usage{
student_t(df, loc, scale)
}
\arguments{
\item{df}{Degrees of freedom parameter.}

\item{loc}{Location parameter.}

\item{scale}{Scale parameter.}
}
\description{
Student-t Distrubtion
}
