% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trans_vars}
\alias{trans_vars}
\title{trans_vars}
\usage{
trans_vars(eqn, map)
}
\arguments{
\item{eqn}{The return value from \code{clean_f()$equations}}

\item{map}{A list that defines the mapping from what the user has declared
(as the states, parameters, and left-hand-side) and the Stan equivalents.}
}
\value{
A character vector that contains the ODE equations in Stan
language syntax.
}
\description{
translate state and parameter variables from R to Stan
}
