% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_ode_generate.R
\name{stan_ode_generate}
\alias{stan_ode_generate}
\title{stan_ode_generate}
\usage{
stan_ode_generate(obj, has_events)
}
\arguments{
\item{obj}{Output of \code{stan_lines()}.}
}
\value{
A list with the stan model to use in \code{rstan::stan()} and more
visually appealing version to be used with \code{cat()}.
}
\description{
Generates a stan program using the ODE system generated by
\code{stanode::stan_lines()}.
}
