% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanode.R
\name{stanode}
\alias{stanode}
\title{stanode}
\usage{
stanode(obj)
}
\arguments{
\item{state_nms}{State names given by the user.}
}
\value{
A list that contains information on the simulation/fitted Stan program.
}
\description{
Create the output for the user after simulating/fitting the Stan program..
}
